Some clean-up and improvements.
