+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.UnitTests/ActionTestsConstants.cs
@@ -2,10 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Diagnostics.Monitoring.WebApi;
-using System;
-using System.Collections.Generic;
-
 namespace Microsoft.Diagnostics.Monitoring.Tool.UnitTests
 {
     internal class ActionTestsConstants
