+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.UnitTests/DiagnosticPortConfigurations/AllListen.txt
@@ -1,0 +1,1 @@
+"DiagnosticPort": "SimplifiedDiagnosticPort"
