+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.UnitTests/DiagnosticPortTestsHelper.cs
@@ -2,117 +2,22 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Diagnostics.Monitoring.Options;
 using Microsoft.Diagnostics.Monitoring.TestCommon;
-using Microsoft.Diagnostics.Monitoring.WebApi.Models;
 using Microsoft.Diagnostics.Tools.Monitor;
-using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Actions;
-using Microsoft.Diagnostics.Tools.Monitor.CollectionRules.Options;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Options;
-using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-using Xunit;
 using Xunit.Abstractions;
 
 namespace Microsoft.Diagnostics.Monitoring.Tool.UnitTests
 {
-    internal static class ActionTestsHelper
+    internal static class DiagnosticPortTestsHelper
     {
-        public static TargetFrameworkMoniker[] tfmsToTest = new TargetFrameworkMoniker[] { TargetFrameworkMoniker.Net50, TargetFrameworkMoniker.Net60 };
-        public static TargetFrameworkMoniker[] tfms6PlusToTest = new TargetFrameworkMoniker[] { TargetFrameworkMoniker.Net60 };
-
-        public static IEnumerable<object[]> GetTfms()
+        public static IEnumerable<object[]> GetFileNamesAndEnvironmentVariables()
         {
-            foreach (TargetFrameworkMoniker tfm in tfmsToTest)
-            {
-                yield return new object[] { tfm };
-            }
-        }
-
-        public static IEnumerable<object[]> Get6PlusTfms()
-        {
-            foreach (TargetFrameworkMoniker tfm in tfms6PlusToTest)
-            {
-                yield return new object[] { tfm };
-            }
-        }
-
-        public static IEnumerable<object[]> GetTfmsAndTraceProfiles()
-        {
-            foreach (TargetFrameworkMoniker tfm in tfmsToTest)
-            {
-                yield return new object[] { tfm, TraceProfile.Logs };
-                yield return new object[] { tfm, TraceProfile.Metrics };
-                yield return new object[] { tfm, TraceProfile.Http };
-                yield return new object[] { tfm, TraceProfile.Cpu };
-            }
-        }
-
-        public static IEnumerable<object[]> GetTfmsAndDumpTypes()
-        {
-            foreach (TargetFrameworkMoniker tfm in tfmsToTest)
-            {
-                yield return new object[] { tfm, DumpType.Full };
-                // Capturing non-full dumps via diagnostic command works inconsistently
-                // on Alpine for .NET 5 and lower (the dump command will return successfully, but)
-                // the dump file will not exist). Only test other dump types on .NET 6+
-                if (!DistroInformation.IsAlpineLinux || tfm == TargetFrameworkMoniker.Net60)
-                {
-                    yield return new object[] { tfm, DumpType.WithHeap };
-                    yield return new object[] { tfm, DumpType.Triage };
-                    yield return new object[] { tfm, DumpType.Mini };
-                }
-            }
-        }
-
-        public static IEnumerable<object[]> GetTfmsAndLogFormat()
-        {
-            foreach (TargetFrameworkMoniker tfm in tfmsToTest)
-            {
-                yield return new object[] { tfm, LogFormat.NewlineDelimitedJson };
-                yield return new object[] { tfm, LogFormat.JsonSequence };
-            }
-        }
-
-        internal static string ValidateEgressPath(CollectionRuleActionResult result)
-        {
-            Assert.NotNull(result.OutputValues);
-            Assert.True(result.OutputValues.TryGetValue(CollectionRuleActionConstants.EgressPathOutputValueName, out string egressPath));
-            Assert.True(File.Exists(egressPath));
-
-            return egressPath;
-        }
-
-        internal static T GetActionOptions<T>(IHost host, string ruleName, int actionIndex = 0)
-        {
-            IOptionsMonitor<CollectionRuleOptions> ruleOptionsMonitor = host.Services.GetService<IOptionsMonitor<CollectionRuleOptions>>();
-            T options = (T)ruleOptionsMonitor.Get(ruleName).Actions[actionIndex].Settings;
-
-            return options;
-        }
-
-        internal async static Task<CollectionRuleActionResult> ExecuteAndDisposeAsync(ICollectionRuleAction action, TimeSpan timeout)
-        {
-            using CancellationTokenSource cancellationTokenSource = new CancellationTokenSource(timeout);
-
-            CollectionRuleActionResult result;
-            try
-            {
-                await action.StartAsync(cancellationTokenSource.Token);
-
-                result = await action.WaitForCompletionAsync(cancellationTokenSource.Token);
-            }
-            finally
-            {
-                await TestCommon.DisposableHelper.DisposeAsync(action);
-            }
-
-            return result;
+            yield return new object[] { "SimplifiedListen.txt", DiagnosticPortTestsConstants.SimplifiedListen_EnvironmentVariables };
+            yield return new object[] { "FullListen.txt", DiagnosticPortTestsConstants.FullListen_EnvironmentVariables };
+            yield return new object[] { "Connect.txt", DiagnosticPortTestsConstants.Connect_EnvironmentVariables };
+            yield return new object[] { "AllListen.txt", DiagnosticPortTestsConstants.AllListen_EnvironmentVariables };
         }
 
         internal static IHostBuilder GetDiagnosticPortHostBuilder(ITestOutputHelper outputHelper, IDictionary<string, string> diagnosticPortEnvironmentVariables)
