+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.UnitTests/Microsoft.Diagnostics.Monitoring.Tool.UnitTests.csproj
@@ -15,6 +15,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <None Update="DiagnosticPortConfigurations\AllListen.txt">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Update="ExpectedConfigurations\AuthenticationFull.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
